Renamed file and added mobile support
