+++ src/components/index.js
@@ -1,4 +1,4 @@
-export {default as Navbar} from "./navbar";
+export {default as Navbar} from "./Navbar";
 export {default as Homepage} from "./Homepage";
 export {default as Exchanges} from "./Exchanges";
 export {default as Cryptocurrencies} from "./Cryptocurrencies";
